+++ src/init.c
@@ -53,8 +53,10 @@ extern SEXP capture_arg(SEXP, SEXP);
 extern SEXP rlang_capturearg(SEXP, SEXP, SEXP, SEXP);
 extern SEXP rlang_capturedots(SEXP, SEXP, SEXP, SEXP);
 extern SEXP r_new_language_(SEXP, SEXP);
+extern SEXP rlang_cnd_abort(SEXP, SEXP);
+extern SEXP rlang_cnd_inform(SEXP, SEXP);
 extern SEXP rlang_cnd_signal(SEXP, SEXP);
-extern SEXP rlang_cnd_signal_error(SEXP, SEXP);
+extern SEXP rlang_cnd_warn(SEXP, SEXP);
 
 // For unit tests
 extern SEXP chr_prepend(SEXP, SEXP);
@@ -105,8 +107,10 @@ static const R_CallMethodDef call_entries[] = {
   {"rlang_unescape_character",  (DL_FUNC) &rlang_unescape_character, 1},
   {"rlang_zap_attrs",           (DL_FUNC) &rlang_zap_attrs, 1},
   {"r_new_language",            (DL_FUNC) &r_new_language_, 2},
+  {"rlang_cnd_abort",           (DL_FUNC) &rlang_cnd_abort, 2},
+  {"rlang_cnd_inform",          (DL_FUNC) &rlang_cnd_inform, 2},
   {"rlang_cnd_signal",          (DL_FUNC) &rlang_cnd_signal, 2},
-  {"rlang_cnd_signal_error",    (DL_FUNC) &rlang_cnd_signal_error, 2},
+  {"rlang_cnd_warn",            (DL_FUNC) &rlang_cnd_warn, 2},
   {"rlang_test_chr_prepend",    (DL_FUNC) &chr_prepend, 2},
   {"rlang_test_chr_append",     (DL_FUNC) &chr_append, 2},
   {"rlang_test_r_warn",         (DL_FUNC) &rlang_test_r_warn, 1},