+++ R/state.R
@@ -9,26 +9,26 @@
 #' * `with_options()` changes options while an expression is
 #'   evaluated. Options are restored when the expression returns.
 #'
-#' * `poke_options()` changes options permanently.
+#' * `push_options()` adds or changes options permanently.
 #'
 #' * `peek_option()` and `peek_options()` return option values. The
 #'   former returns the option directly while the latter returns a
 #'   list.
 #'
-#' @param ... For `scoped_options()` and `poke_options()`, named
+#' @param ... For `scoped_options()` and `push_options()`, named
 #'   values defining new option values. For `peek_options()`, strings
 #'   or character vectors of option names.
 #' @param .frame The environment of a stack frame which defines the
 #'   scope of the temporary options. When the frame returns, the
 #'   options are set back to their original values.
-#' @return For `scoped_options()` and `poke_options()`, the old option
+#' @return For `scoped_options()` and `push_options()`, the old option
 #'   values. `peek_option()` returns the current value of an option
 #'   while the plural `peek_options()` returns a list of current
 #'   option values.
 #' @export
 #' @examples
 #' # Store and retrieve a global option:
-#' poke_options(my_option = 10)
+#' push_options(my_option = 10)
 #' peek_option("my_option")
 #'
 #' # Change the option temporarily:
@@ -66,7 +66,7 @@ with_options <- function(.expr, ...) {
 }
 #' @rdname scoped_options
 #' @export
-poke_options <- function(...) {
+push_options <- function(...) {
   options(dots_list(...))
 }
 #' @rdname scoped_options