+++ NEWS.md
@@ -67,7 +67,7 @@
   bindings in an environment.
 
 * New API for changing global options: `peek_options()` and
-  `peek_option()` examine options; `poke_options()` changes options
+  `peek_option()` examine options; `push_options()` changes options
   indefinitely; `scoped_options()` and `with_options()` change options
   temporarily.